refactor & change sigma_max
