+++ fastvideo/training/wan_rl_training_pipeline.py
@@ -1,12 +1,10 @@
 # SPDX-License-Identifier: Apache-2.0
 import sys
-from copy import deepcopy
 
 from fastvideo.fastvideo_args import FastVideoArgs, TrainingArgs
 from fastvideo.logger import init_logger
 from fastvideo.models.schedulers.scheduling_unipc_multistep import (
     UniPCMultistepScheduler)
-from fastvideo.pipelines.basic.wan.wan_pipeline import WanPipeline
 from fastvideo.training.rl.rl_pipeline import RLPipeline
 from fastvideo.utils import is_vsa_available
 
@@ -37,54 +35,10 @@ class WanRLTrainingPipeline(RLPipeline):
         """
         pass
 
-    def _create_inference_pipeline(self, training_args: TrainingArgs,
-                                   dit_cpu_offload: bool):
-        args_copy = deepcopy(training_args)
-        args_copy.inference_mode = True
-        loaded_modules = {
-            "transformer": self.get_module("transformer"),
-        }
-        transformer_2 = self.get_module("transformer_2", None)
-        if transformer_2 is not None:
-            loaded_modules["transformer_2"] = transformer_2
-        text_encoder = self.get_module("text_encoder", None)
-        if text_encoder is not None:
-            loaded_modules["text_encoder"] = text_encoder
-        tokenizer = self.get_module("tokenizer", None)
-        if tokenizer is not None:
-            loaded_modules["tokenizer"] = tokenizer
-        vae = self.get_module("vae", None)
-        if vae is not None:
-            loaded_modules["vae"] = vae
-        # Use UniPCMultistepScheduler for RL sampling to match flow_grpo
-        scheduler = self.get_module("scheduler", None)
-        if scheduler is not None:
-            loaded_modules["scheduler"] = scheduler
-
-        pipeline = WanPipeline.from_pretrained(
-            training_args.model_path,
-            args=args_copy,  # type: ignore
-            inference_mode=True,
-            loaded_modules=loaded_modules,
-            tp_size=training_args.tp_size,
-            sp_size=training_args.sp_size,
-            num_gpus=training_args.num_gpus,
-            pin_cpu_memory=training_args.pin_cpu_memory,
-            dit_cpu_offload=dit_cpu_offload)
-        # Override scheduler to use UniPCMultistepScheduler
-        if scheduler is not None:
-            pipeline.modules["scheduler"] = scheduler
-        return pipeline
-
     def initialize_validation_pipeline(self, training_args: TrainingArgs):
         logger.info("Initializing validation pipeline...")
         self.validation_pipeline = self._create_inference_pipeline(
             training_args, dit_cpu_offload=True)
-
-    def _build_sampling_pipeline(self, training_args: TrainingArgs):
-        return self._create_inference_pipeline(training_args,
-                                               dit_cpu_offload=False)
-
 
 def main(args) -> None:
     logger.info("Starting RL training pipeline...")
